update error handling
